+++ frontend/sense-and-art-frontend/src/ImagePicker/CloseButton.css
