+++ frontend/sense-and-art-frontend/src/ImagePicker/CloseButton.js
