+++ frontend/sense-and-art-frontend/src/ImagePicker/Image.js
