+++ frontend/sense-and-art-frontend/src/ImagePicker/ImageGrid.css
