+++ frontend/sense-and-art-frontend/src/ImagePicker/ImageGuide.css
