+++ frontend/sense-and-art-frontend/src/ImagePicker/SelectImageButton.css
