+++ frontend/sense-and-art-frontend/src/ImagePicker/ZoomedImage.css
