+++ frontend/sense-and-art-frontend/src/ImagePicker/ZoomedImagePage.js
